requirements.txt
# Core plotting and numerical libraries
matplotlib>=3.5.0
numpy>=1.21.0

# Image processing
Pillow>=9.0.0
opencv-python>=4.5.0

# Deep Learning & Object Detection
ultralytics>=8.0.0
torch>=2.0.0

# Optical Character Recognition
easyocr>=1.7.0
scikit-image>=0.19.0

# Utilities & Configuration
PyYAML>=5.4.0
tqdm>=4.60.0

# Development & Optional
jupyter>=1.0.0
ipython>=7.0.0
